Fix AMP dtype mismatch in SetSpanLoss target assignment
presence.new_zeros() inherits presence's autocast dtype (Half on GPU);
valid_spans/ev_w come from the batch collate and stay Float32, so the
index_put assignment crashed every DDP rank at the first real-GPU training
step with "Index put requires the source and destination dtypes match".
Never reproduced in CPU-only local tests since autocast wasn't engaged
there - caught by the Kaggle GPU smoke test.

+++ src/train/set_losses.py
@@ -105,11 +105,11 @@ class SetSpanLoss:
         for b, (qi, ti) in enumerate(matches):
             if qi.numel() == 0:
                 continue
-            valid_spans = spans_sec[b][valid[b]]
+            valid_spans = spans_sec[b][valid[b]].to(tgt_s.dtype)
             pos[b, qi] = 1.0
             tgt_s[b, qi] = valid_spans[ti, 0]
             tgt_e[b, qi] = valid_spans[ti, 1]
-            bw[b, qi] = ev_w[b]
+            bw[b, qi] = ev_w[b].to(bw.dtype)
 
         n_pos = pos.sum().clamp(min=1.0)
         pres_w = torch.where(pos > 0.5, torch.ones_like(pos),
